requirements.txt
Flask==2.3.3
line-bot-sdk==3.5.0
python-dotenv==1.0.0
openpyxl==3.1.2
Pillow==10.0.1
pdf2image==1.16.3
pypdf==3.17.1
requests==2.31.0
python-dateutil==2.8.2
pdfplumber==0.10.3
reportlab==4.0.7
numpy==1.24.3
